refactor: DiaryService
