+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/diary/authorization/DiaryOwnerAuthorizationInterceptorTest.java
@@ -1,6 +1,6 @@
 package com.startingblue.fourtooncookie.diary.authorization;
 
-import com.startingblue.fourtooncookie.diary.service.DiaryService;
+import com.startingblue.fourtooncookie.diary.DiaryService;
 import com.startingblue.fourtooncookie.web.interceptor.DiaryOwnerAuthorizationInterceptor;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
