+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/diary/service/DiarySQSMessageListenerTest.java
@@ -1,6 +1,7 @@
 package com.startingblue.fourtooncookie.diary.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.startingblue.fourtooncookie.diary.DiaryService;
 import com.startingblue.fourtooncookie.diary.listener.DiarySQSMessageListener;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
